Move some stuff from socket.rb to socket.cpp
...using the new __bind_method__ macro.
